+++ frontend/src/lib/components/apps/editor/contextPanel/components/OutputHeader.svelte
@@ -93,7 +93,9 @@
 		if (
 			item?.data.type == 'aggridcomponent' ||
 			item?.data.type == 'aggridcomponentee' ||
-			item?.data.type == 'dbexplorercomponent'
+			item?.data.type == 'dbexplorercomponent' ||
+			item?.data.type == 'aggridinfinitecomponent' ||
+			item?.data.type == 'aggridinfinitecomponentee'
 		) {
 			for (let c of item.data.actions ?? []) {
 				let old = c.id
@@ -126,7 +128,9 @@
 		if (
 			(data.type == 'aggridcomponent' ||
 				data.type == 'aggridcomponentee' ||
-				data.type == 'dbexplorercomponent') &&
+				data.type == 'dbexplorercomponent' ||
+				data.type == 'aggridinfinitecomponent' ||
+				data.type == 'aggridinfinitecomponentee') &&
 			Array.isArray(data.actions)
 		) {
 			for (let c of data.actions) {